pyproject.toml
[project]
name = "agent-tutorials-zh"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.2.0",
    "ollama>=0.4.8",
    "loguru>=0.7.3",
]
